+++ R/predict.R
@@ -264,7 +264,8 @@ normTable <- function(A,
 #' # generate a norm table for the raw value range from 0 to 28 for month 7 of grade 3
 #' normData <- prepareData()
 #' m <- bestModel(data=normData)
-#' table <- rawTable(3 + 7/12, m, 0, 28, precision=.01)
+#' table <- rawTable(3 + 7/12, m, minRaw = 0, maxRaw = 28,
+#'                   minNorm = 25, maxNorm = 75, precision=.01)
 #' @export
 rawTable <- function(A,
                      model,